Got chromosomes working, though there's really only one trait to modify right now. Started working on combat, but it's not implemented yet.
git-svn-id: http://pest-control.googlecode.com/svn/trunk@25 c1a93738-a89d-add3-e846-535381a29ce9

+++ pest-control/pest-control/game/LivingThing.cs
@@ -10,6 +10,6 @@ namespace pest_control
 {
     public abstract class LivingThing : MovingThing
     {
-        protected int health;
+        protected double health;
     }
 }
